django.5-postgres/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-base64-binaryfield==1.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
pillow==12.0.0
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.2
sqlparse==0.5.3
